refacto(http-handler): modularize tree files parser
